refactor dict module
