--- vortex-array/src/arrays/dict/arrow.rs
@@ -1,19 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: Copyright the Vortex contributors
-
-use arrow_array::types::ArrowDictionaryKeyType;
-use arrow_array::{AnyDictionaryArray, DictionaryArray};
-
-use super::DictArray;
-use crate::ArrayRef;
-use crate::arrow::FromArrowArray;
-
-impl<K: ArrowDictionaryKeyType> FromArrowArray<&DictionaryArray<K>> for DictArray {
-    fn from_arrow(array: &DictionaryArray<K>, nullable: bool) -> Self {
-        let keys = AnyDictionaryArray::keys(array);
-        let keys = ArrayRef::from_arrow(keys, keys.is_nullable());
-        let values = ArrayRef::from_arrow(array.values().as_ref(), nullable);
-        // SAFETY: we assume that Arrow has checked the invariants on construction
-        unsafe { DictArray::new_unchecked(keys, values) }
-    }
-}
