+++ vortex-array/src/arrays/dict/vtable/encode.rs
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_error::VortexResult;
+
+use super::DictVTable;
+use crate::Canonical;
+use crate::arrays::dict::DictArray;
+use crate::builders::dict::dict_encode;
+use crate::vtable::EncodeVTable;
+
+impl EncodeVTable<DictVTable> for DictVTable {
+    fn encode(
+        _vtable: &DictVTable,
+        canonical: &Canonical,
+        _like: Option<&DictArray>,
+    ) -> VortexResult<Option<DictArray>> {
+        Ok(Some(dict_encode(canonical.as_ref())?))
+    }
+}
