+++ vortex-array/src/arrays/dict/vtable/visitor.rs
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use super::DictVTable;
+use crate::arrays::dict::DictArray;
+use crate::vtable::VisitorVTable;
+use crate::{ArrayBufferVisitor, ArrayChildVisitor};
+
+impl VisitorVTable<DictVTable> for DictVTable {
+    fn visit_buffers(_array: &DictArray, _visitor: &mut dyn ArrayBufferVisitor) {}
+
+    fn visit_children(array: &DictArray, visitor: &mut dyn ArrayChildVisitor) {
+        visitor.visit_child("codes", array.codes());
+        visitor.visit_child("values", array.values());
+    }
+}
